+++ packages/http-client/src/http-client.ts
@@ -32,6 +32,7 @@ import {
   HttpRequestConfigWithOrigin,
   HttpRequestOptions,
   HttpResponse,
+  isHttpRequestConfigWithOrigin,
   OriginOptions,
   OriginOptionsInternal,
   ParameterEncoder
@@ -111,7 +112,7 @@ export function execute<ReturnT>(executeFn: ExecuteHttpRequestFn<ReturnT>) {
     const destinationRequestConfig = await buildHttpRequest(
       resolvedDestination
     );
-    logCustomHeadersWarning(requestConfig.headers?.custom);
+    logCustomHeadersWarning(requestConfig.headers);
     const request = await buildRequestWithMergedHeadersAndQueryParameters(
       requestConfig,
       resolvedDestination,
@@ -121,6 +122,31 @@ export function execute<ReturnT>(executeFn: ExecuteHttpRequestFn<ReturnT>) {
     request.headers = await addCsrfTokenToHeader(destination, request, options);
     logRequestInformation(request);
     return executeFn(request);
+  };
+}
+
+/**
+ * Build an [[HttpRequestConfigWithOrigin]] from a given [[HttpRequestConfigWithOrigin]] or [[HttpRequestConfig]]
+ * @param requestConfig - The given [[HttpRequestConfigWithOrigin]] or [[HttpRequestConfig]]
+ * @returns The resulting [[HttpRequestConfigWithOrigin]]
+ * @internal
+ */
+export function buildHttpRequestConfigWithOrigin(
+  requestConfig: HttpRequestConfigWithOrigin | HttpRequestConfig
+): HttpRequestConfigWithOrigin {
+  if (isHttpRequestConfigWithOrigin(requestConfig)) {
+    return requestConfig;
+  }
+  return {
+    ...requestConfig,
+    headers: {
+      requestConfig: {},
+      ...(requestConfig.headers && { custom: requestConfig.headers })
+    },
+    params: {
+      requestConfig: {},
+      ...(requestConfig.params && { custom: requestConfig.params })
+    }
   };
 }
 
@@ -273,16 +299,25 @@ function splitRequestConfig(requestConfig: HttpRequestConfigWithOrigin): {
   };
 }
 
-function logCustomHeadersWarning(customHeaders?: Record<string, string>) {
-  if (customHeaders) {
-    logger.debug(
-      `The following custom headers will overwrite headers created by the SDK, if they use the same key:\n${Object.keys(
-        customHeaders
-      )
-        .map(key => `  - "${key}"`)
-        .join('\n')}
-If the parameters from multiple origins use the same key, the priority is 1. Custom, 2. Destination, 3. Internal.`
+function logCustomHeadersWarning(headers?: OriginOptions) {
+  if (!headers) {
+    return;
+  }
+  const customHeaders = headers.custom;
+  const requestConfigHeaders = headers.requestConfig;
+  if (customHeaders && requestConfigHeaders) {
+    const headerKeysToBeOverwritten = Object.keys(customHeaders).filter(
+      customHeaderKey =>
+        Object.keys(requestConfigHeaders).includes(customHeaderKey)
     );
+    if (headerKeysToBeOverwritten.length) {
+      logger.debug(
+        `The following custom headers will overwrite headers created by the SDK, if they use the same key:\n${headerKeysToBeOverwritten
+          .map(key => `  - "${key}"`)
+          .join('\n')}
+If the parameters from multiple origins use the same key, the priority is 1. Custom, 2. Destination, 3. Internal.`
+      );
+    }
   }
 }
 
@@ -302,17 +337,62 @@ function logRequestInformation(request: HttpRequestConfig) {
 }
 
 /**
+ * Builds a [[DestinationHttpRequestConfig]] for the given destination, merges it into the given `requestConfig`
+ * and executes it (using Axios).
+ * The overload, that accepts [[HttpRequestConfigWithOrigin]] as a parameter, is deprecated and replaced the function [[executeHttpRequestWithOrigin]].
+ * @param destination - A destination or a destination name and a JWT.
+ * @param requestConfig - Any object representing an HTTP request.
+ * @param options - An [[HttpRequestOptions]] of the HTTP request for configuring e.g., CSRF token delegation. By default, the SDK will fetch the CSRF token.
+ * @returns A promise resolving to an [[HttpResponse]].
+ */
+export function executeHttpRequest<T extends HttpRequestConfig>(
+  destination: DestinationOrFetchOptions,
+  requestConfig: T,
+  options?: HttpRequestOptions
+): Promise<HttpResponse>;
+/**
+ * @deprecated This overload is replaced by the function [[executeHttpRequestWithOrigin]].
+ */
+export function executeHttpRequest<T extends HttpRequestConfigWithOrigin>(
+  destination: DestinationOrFetchOptions,
+  requestConfig: T,
+  options?: HttpRequestOptions
+): Promise<HttpResponse>;
+// eslint-disable-next-line jsdoc/require-jsdoc
+export function executeHttpRequest<
+  T extends HttpRequestConfig | HttpRequestConfigWithOrigin
+>(
+  destination: DestinationOrFetchOptions,
+  requestConfig: T,
+  options?: HttpRequestOptions
+): Promise<HttpResponse> {
+  // eslint-disable-next-line jsdoc/require-jsdoc
+  const requestConfigWithOrigin =
+    buildHttpRequestConfigWithOrigin(requestConfig);
+  return execute(executeWithAxios)(
+    destination,
+    requestConfigWithOrigin,
+    options
+  );
+}
+
+/**
  * Builds a [[DestinationHttpRequestConfig]] for the given destination, merges it into the given [[HttpRequestConfigWithOrigin]]
  * and executes it (using Axios).
  * The [[HttpRequestConfigWithOrigin]] supports defining header options and query parameter options with origins.
- * When reaching conflicts, values with high priorities are chosen.
- * The priorities are defined in the [[origins]].
+ * Equally named headers and query parameters are prioritized in the following order:
+ * 1. `custom`
+ * 2. Destination related headers/query parameters
+ * 3. `requestConfig`.
  * @param destination - A destination or a destination name and a JWT.
  * @param requestConfig - Any object representing an HTTP request.
- * @param options - An [[HttpRequestOptions]] of the http request for configuring e.g., CSRF token delegation. By default, the SDK will not fetch the CSRF token.
+ * @param options - An [[HttpRequestOptions]] of the HTTP request for configuring e.g., CSRF token delegation. By default, the SDK will fetch the CSRF token.
  * @returns A promise resolving to an [[HttpResponse]].
+ * @see https://sap.github.io/cloud-sdk/docs/js/features/connectivity/query-parameters
  */
-export function executeHttpRequest<T extends HttpRequestConfigWithOrigin>(
+export function executeHttpRequestWithOrigin<
+  T extends HttpRequestConfigWithOrigin
+>(
   destination: DestinationOrFetchOptions,
   requestConfig: T,
   options?: HttpRequestOptions