+++ test-packages/type-tests/test/v2/filter-functions.spec.ts
@@ -1,6 +1,4 @@
-import {
-  testService
-} from '@sap-cloud-sdk/test-services-odata-v2/test-service';
+import { testService } from '@sap-cloud-sdk/test-services-odata-v2/test-service';
 import {
   filterFunctions,
   length,