+++ test-packages/type-tests/test/v2/filter.spec.ts
@@ -40,9 +40,11 @@ testEntitySchema.TO_SINGLE_LINK.filter(
 );
 
 // $ExpectType FilterList<TestEntity<DeSerializers<string, boolean, number, BigNumber, number, number, number, number, BigNumber, string, number, number, string, any, Moment, Moment, Time>>, DeSerializers<string, boolean, number, BigNumber, number, number, number, number, BigNumber, string, number, number, string, any, Moment, Moment, Time>>
-and(testEntitySchema.TO_SINGLE_LINK.filter(
- testEntitySingleLinkSchema.STRING_PROPERTY.equals('test')
-));
+and(
+  testEntitySchema.TO_SINGLE_LINK.filter(
+    testEntitySingleLinkSchema.STRING_PROPERTY.equals('test')
+  )
+);
 
 // $ExpectType Filter<TestEntity<DeSerializers<any, any, any, any, any, any, any, any, any, any, any, any, any, any, any, any, any>>, DeSerializers<string, boolean, number, BigNumber, number, number, number, number, BigNumber, string, number, number, string, any, Moment, Moment, Time>, string>
 testEntitySchema.COMPLEX_TYPE_PROPERTY.stringProperty.equals('test');