+++ packages/openapi/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-cloud-sdk/openapi
 
+## 2.5.0
+
+### Patch Changes
+
+- Updated dependencies [89f1c423]
+- Updated dependencies [9481ec69]
+  - @sap-cloud-sdk/util@2.5.0
+  - @sap-cloud-sdk/http-client@2.5.0
+  - @sap-cloud-sdk/connectivity@2.5.0
+
 ## 2.4.0
 
 ### Patch Changes